feat: modularize profile feature domain
