+++ src/features/profile/components/index.ts
@@ -1,0 +1,4 @@
+export { ProfileEditForm } from './ProfileEditForm';
+export type { ProfileEditFormProps } from './ProfileEditForm';
+export { PasswordChangeForm } from './PasswordChangeForm';
+export type { PasswordChangeFormProps } from './PasswordChangeForm';
