+++ src/features/profile/index.ts
@@ -1,0 +1,4 @@
+export * from './components';
+export * from './views';
+export * from './types/profile';
+export { default as UserProfileSystem } from '@/components/UserProfileSystem';
