+++ src/features/profile/views/index.ts
@@ -1,0 +1,3 @@
+export { ArtistMyPage } from './ArtistMyPage';
+export { FanMyPage } from './FanMyPage';
+export { AdminMyPage } from './AdminMyPage';
